+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Fix
 
+- BigQuery: the JWT token auth method is now supported in the status check.
 - HTTP: allow connector to be instanciated without passing positional arguments to auth.
 
 ## [6.3.0] 2024-06-21